chore(eslint): upgrade eslint-plugin-tailwind
