+++ packages/eslint-config-bases/src/bases/index.cjs
@@ -4,8 +4,8 @@ module.exports = {
   react: require('./react'),
   regexp: require('./regexp'),
   reactTestingLibrary: require('./rtl'),
-  sonar: require('./sonar.mjs'),
+  sonar: require('./sonar.js'),
   storybook: require('./storybook'),
   typescript: require('./typescript.mjs'),
-  tailwind: require('./tailwind'),
+  tailwind: require('./tailwind.mjs'),
 };
