+++ packages/tailwind-constructor/package.json
@@ -22,6 +22,6 @@
     "autoprefixer": "10.4.14",
     "postcss": "8.4.27",
     "postcss-import": "^14.0.2",
-    "tailwindcss": "3.4.3"
+    "tailwindcss": "3.4.11"
   }
 }
